pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "highrise-bot-sdk>=24.1.0",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "spotipy>=2.25.1",
    "youtube-dl>=2021.12.17",
]
